feat(moderators): add moderators cards (#32)
* feat(moderators): add some moderators cards

The UI for the horizontal cards isn't finished on the PR #21. So the moderators code will change in the future.

Also, some moderators are missing since I couldn't find their GitHub on the organization. This needs to be added later as well.

Issue: #6

* feat(moderators): add missing moderators #32

Issue: #6

* fix(moderators): moderator card width in small screens

When a moderator card doesn't have description or it's too short, it won't occupy the whole width like the other cards.

Issue: #6

* refactor(moderators): update moderators card to UI on #21

Issue: #6

* fix(moderators): add a white background and header

Removed "mx-auto" class since it is not necessary to make the div centered.

* fix(moderators): contrast of community label